Share handleArray and handleValdiator root functions
