+++ beacon-chain/state/stateutil/BUILD.bazel
@@ -4,6 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "array_root.go",
         "block_header_root.go",
         "eth1_root.go",
         "pending_attestation_root.go",
@@ -47,6 +48,7 @@ go_test(
         "state_root_test.go",
         "stateutil_test.go",
         "trie_helpers_test.go",
+        "validator_root_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
